# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "go-aliyunmc",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
...
        "/user": {
            "get": {
                "description": "根据传入的凭证返回该用户的非敏感信息，例如用户名、ID、创建时间、权限组等信息。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "获取当前已登入用户信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.DataResp-store_User"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            },
            "post": {
                "description": "根据提供的用户名和密码创建一个新的用户。用户名是唯一的，如果与现有用户重复则返回409。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "创建用户请求体",
                        "name": "createuserrequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            }
        },
        "/user/{userId}": {
            "delete": {
                "description": "删除某个用户，使得该用户名以后无法再次登录。此接口仅适用于用户主动注销，如果请求参数携带的用户ID与凭证中的用户ID不匹配，返回403",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "注销用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "删除目标用户ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            },
            "patch": {
                "description": "更新用户在数据库中存储的各个字段。目前仅支持更新密码。",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "目标用户ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新请求体",
                        "name": "updateuserrequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "helpers.DataResp-store_User": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data 为该结构中的数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/store.User"
                        }
                    ]
                }
            }
        },
...
        "helpers.ErrorResp": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Details 为错误信息",
                    "type": "string"
                }
            }
        },
...
        "store.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "users.CreateUserRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Password 是要创建用户的密码明文",
                    "type": "string"
                },
                "username": {
                    "description": "Username 是要创建用户的用户名，不可与其他用户重复",
                    "type": "string"
                }
            }
        },
        "users.UpdateUserRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "description": "Password 是新密码的明文",
                    "type": "string"
                }
            }
        }
    }
}